microservice-app/notification-service/requirements.txt
fastapi
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
prometheus-fastapi-instrumentator
opentelemetry-exporter-zipkin
protobuf<4.0.0,>=3.20.0
